pyproject.toml
[project]
name = "daily-digest"
version = "0.1.0"
description = "Application that scrapes and summarizes the most recent news, sends trend emails"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
      "beautifulsoup4==4.12.2",
      "requests==2.31.0",
      "pandas>=2.1.4",
      "numpy>=1.26.0",
      "scikit-learn>=1.3.2",
      "nltk>=3.8.1",
      "fastapi>=0.104.1",
      "uvicorn>=0.24.0",
      "apscheduler>=3.10.4",
      "pyyaml>=6.0.1",
      "textblob>=0.17.1",
      "vaderSentiment>=3.3.2",
      "pytest>=7.4.3",
      "pytest-asyncio>=0.21.0",
      "pytest-cov>=4.1.0",
      "lxml>=4.9.3",
      "jinja2>=3.1.2",
      "aiosmtplib>=3.0.0",
      "email-validator>=2.0.0",
      "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
